fix missing updates on the changeover to int based body lengths

+++ artio-core/src/main/java/uk/co/real_logic/artio/engine/PossDupEnabler.java
@@ -369,8 +369,8 @@ public class PossDupEnabler
     {
         final int frameBodyLengthOffset =
             claimOffset + MessageHeaderDecoder.ENCODED_LENGTH + FixMessageDecoder.BLOCK_LENGTH + metaDataAdjustment;
-        final short frameBodyLength = (short)(messageLength + lengthDelta);
-        claimBuffer.putShort(frameBodyLengthOffset, frameBodyLength, LITTLE_ENDIAN);
+        final int frameBodyLength = messageLength + lengthDelta;
+        claimBuffer.putInt(frameBodyLengthOffset, frameBodyLength, LITTLE_ENDIAN);
     }
 
     private void updateBodyLengthAndChecksum(
